Minor test changes (Part 11)
