+++ docs/04-devops/01-notes.md.txt
@@ -20,6 +20,7 @@ export JAVA_HOME=~/jdk-21.0.9/
 ./gradlew :project-build:postgre-repo:clean
 ./gradlew :project-build:postgre-repo:update --info
 ./gradlew :project-build:postgre-repo:test --info
+./gradlew :project-build:app-string:test --info -Puse.db.server=true -Puse.db.container=true
 
 - запуск докер контейнер с Postgre и контроль работы. Ссылка на docker image - https://hub.docker.com/_/postgres
     docker run --name some-postgres -e POSTGRES_PASSWORD=mysecretpassword -d -p 5432:5432 postgres:15.4
