+++ project-build/gradle.properties
@@ -1,3 +1,5 @@
 kotlin.code.style=official
 kotlinVersion=2.1.21
-jooq.enabled=false 
+jooq.enabled=false
+use.db.server=false
+use.db.container=false
